Potential fix for index bundle iterators
A first attempt at trying to get index bundle  iterators to respect the
requested number of results passed in to nextPages when using a crawl mix. The
union iterator still appears to have problems because it sets results_per_block
to the sum of the results_per_block values for each iterator it contains.

Signed-off-by: Chris Pollett <chris@pollett.org>