Creating another patch With all suggested changes, Please see summary below (1) The single feed and category icons look different in size. This might be an optical illusion caused by differences in padding in the two icons, but is still disconcerting. The category icon should not be bar, two indented bars, bar, indented bar. Instead, just use bar indented bar, bar indented bar. - Fixed icons as per your suggestion. (2) The single feed and category buttons only work when using the admin (non-collapsed) controller. They don't show up at all when using the group (collapsed) controller. - This is fixed, also When moving from expanded controller to collapsed controller, view is also retained. (3) Category view is not doing sufficient isset checks. It generates Notices like: NOTICE: Undefined index: SUBTITLE at line 104 in /Applications/MAMP/htdocs/git/yioop/views/elements/groupfeed_element.php
NOTICE: Undefined index: SUBTITLE at line 122 in /Applications/MAMP/htdocs/git/yioop/views/elements/groupfeed_element.php

- These errors and another on the same page are all fixed byt checking using isset()

(4) In the wiki, I don't see what the Preview button does which is useful. When I click either Save or Preview it says that the page is saved. So it does not seem to be previewing changes. For the extra-button I would really like a strong use-case for why it is needed.
- When you click on Preview, that doesnt save the wiki content. I am not sure why this happened when you tried. I tried again multiple times and clicking on Preview just shows the preview. The idea behing having a seperate preview button is to save additional POST call to the server, Again, this is somethign I found very handy on Media wiki.

(5) You are also adding a lot of code for the help wiki, but no way the end user can test it. This needs to be testable before the patch goes live. Have at least one place where help has been added so I can test this new functionality.
- I added a sample help button on manageGroups action -  ?c=admin&YIOOP_TOKEN=obbrFXEgrts|1413522871&a=manageGroups. Please let me know If you need the buttons rendered anywhere else.

Signed-off-by: Chris Pollett <chris@pollett.org>