- Making changes to "start new thread" following feedback from professor.
(2) Also, the title attribute tooltip Start New Thread should maybe say Start New Thread in [Name of Group]. Then when you click it the form that opens should have as its title Start New Thread in [Name of Group]. Right now the title of the form is Start New Thread.
(3) Start New Thread seems somewhat awkwardly aligned as it appears to the right of the person's icon rather than beneath it. On the other hand, the rest of the form is beneath the icon. Try to reformat this.

Signed-off-by: epinapala <epinapala@ebay.com>
Signed-off-by: Chris Pollett <chris@pollett.org>